+++ common/src/main/java/top/vmctcn/vmtranslationupdate/config/ModConfigs.java
@@ -8,8 +8,9 @@ import top.vmctcn.vmtranslationupdate.respack.ResPackSource;
 
 @Config(name = VMTranslationUpdate.MOD_ID)
 public class ModConfigs implements ConfigData {
+    public boolean testMode = false;
+
     public boolean autoSwitchLanguage = true;
-    public String switchLanguage = "zh_cn";
 
     public boolean autoDownloadVMTranslationPack = false;
     @ConfigEntry.Gui.Tooltip
@@ -17,10 +18,6 @@ public class ModConfigs implements ConfigData {
     public ResPackSource translationPackSource = ResPackSource.GITEE;
 
     public boolean checkModPackTranslationUpdate = true;
-    public String modPackTranslationUpdateCheckUrl = "https://gitee.com/Wulian233/vmtu/raw/main/update/example.txt";
-    public String modPackTranslationUrl = "https://vmct-cn.top/modpacks/example/";
-
-    public String modPackTranslationVersion = "1.0.0";
 
     public boolean i18nUpdateModCheck = true;
     public boolean vaultPatcherCheck = false;