+++ common/src/main/java/top/vmctcn/vmtranslationupdate/modpack/VersionChecker.java
@@ -1,4 +1,4 @@
-package top.vmctcn.vmtranslationupdate.util;
+package top.vmctcn.vmtranslationupdate.modpack;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
@@ -6,10 +6,10 @@ import java.net.URL;
 import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
 
-public class VersionCheckUtil {
+public class VersionChecker {
     public static String getOnlineVersion() {
         try {
-            URL url = new URL(ModConfigUtil.getConfig().modPackTranslationUpdateCheckUrl);
+            URL url = new URL(ModpackInfoReader.getModpackInfo().getModpack().getTranslation().getUpdateCheckUrl());
             URLConnection connection = url.openConnection();
 
             String userAgent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/109.0.5414.120 Safari/537.36 MCMod/VmUpdate";