+++ common/src/main/java/top/vmctcn/vmtranslationupdate/respack/GameOptionsSetter.java
@@ -1,7 +1,8 @@
 package top.vmctcn.vmtranslationupdate.respack;
 
 import top.vmctcn.vmtranslationupdate.VMTranslationUpdate;
-import top.vmctcn.vmtranslationupdate.util.ModConfigUtil;
+import top.vmctcn.vmtranslationupdate.config.ModConfigHelper;
+import top.vmctcn.vmtranslationupdate.modpack.ModpackInfoReader;
 import top.vmctcn.vmtucore.ModPlatform;
 import top.vmctcn.vmtucore.VMTUCore;
 import top.vmctcn.vmtucore.respack.GameOptionsWriter;
@@ -11,19 +12,21 @@ import java.nio.file.Path;
 
 public class GameOptionsSetter {
     public static void init(Path gamePath) {
-        if (ModConfigUtil.getConfig().autoSwitchLanguage && ModConfigUtil.getConfig().switchLanguage != null) {
+        if (ModConfigHelper.getConfig().autoSwitchLanguage && ModpackInfoReader.getModpackInfo().getModpack().getTranslation().getLanguage() != null) {
             try {
-                GameOptionsWriter config = new GameOptionsWriter(gamePath.resolve("options.txt"));
-                config.switchLanguage(ModConfigUtil.getConfig().switchLanguage);
+                GameOptionsWriter writer = new GameOptionsWriter(gamePath.resolve("options.txt"));
+                String lang = ModpackInfoReader.getModpackInfo().getModpack().getTranslation().getLanguage();
+                writer.switchLanguage(LanguageHelper.getFixedLanguage(lang));
             } catch (IOException e) {
                 VMTranslationUpdate.LOGGER.warn("Failed to switch language: ", e);
             }
         }
 
-        if (ModConfigUtil.getConfig().autoDownloadVMTranslationPack) {
+        if (ModConfigHelper.getConfig().autoDownloadVMTranslationPack) {
             String gameVersion = ModPlatform.INSTANCE.getGameVersion();
-            ResPackSource resPackSource = ModConfigUtil.getConfig().translationPackSource;
-            VMTUCore.init(gamePath, gameVersion, "VM汉化组模组汉化包1.20", resPackSource.getUrl());
+            ResPackSource resPackSource = ModConfigHelper.getConfig().translationPackSource;
+            String resPackName = ModpackInfoReader.getModpackInfo().getModpack().getTranslation().getResourcePackName();
+            VMTUCore.init(gamePath, gameVersion, resPackName, resPackSource.getUrl());
         }
     }
 }